get simd runtime detection working
